+++ src/libsrc++/evioSocketChannel.cc
@@ -155,7 +155,7 @@ bool evioSocketChannel::read(uint32_t *myBuf, int length) throw(evioException) {
  *
  * Note:  user MUST free the allocated buffer!
  */
-bool evioSocketChannel::readAlloc(uint32_t **buffer, uint64_t *bufLen) throw(evioException) {
+bool evioSocketChannel::readAlloc(uint32_t **buffer, uint32_t *bufLen) throw(evioException) {
   noCopyBuf=NULL;
   if(handle==0)throw(evioException(0,"evioSocketChannel::readAlloc...0 handle",__FILE__,__FUNCTION__,__LINE__));
 
@@ -182,7 +182,7 @@ bool evioSocketChannel::readAlloc(uint32_t **buffer, uint64_t *bufLen) throw(evi
 bool evioSocketChannel::readNoCopy(void) throw(evioException) {
   if(handle==0)throw(evioException(0,"evioSocketChannel::readNoCopy...0 handle",__FILE__,__FUNCTION__,__LINE__));
 
-  uint64_t bufLen;
+  uint32_t bufLen;
   int stat=evReadNoCopy(handle,&noCopyBuf,&bufLen);
   if(stat==EOF) return(false);
   if(stat!=S_SUCCESS) throw(evioException(stat,"evioSocketChannel::readNoCopy...read error: " + string(evPerror(stat)),
