+++ src/libsrc/evio.h
@@ -82,8 +82,8 @@ int evOpenBuffer(char *buffer, size_t bufLen, char *flags, int *handle);
 int evOpenSocket(int sockFd, char *flags, int *handle);
 
 int evRead(int handle, uint32_t *buffer, size_t size);
-int evReadAlloc(int handle, uint32_t **buffer, uint64_t *buflen);
-int evReadNoCopy(int handle, const uint32_t **buffer, uint64_t *buflen);
+int evReadAlloc(int handle, uint32_t **buffer, uint32_t *buflen);
+int evReadNoCopy(int handle, const uint32_t **buffer, uint32_t *buflen);
 int evReadRandom(int handle, const uint32_t **pEvent, size_t eventNumber);
 int evGetRandomAccessTable(int handle, const uint32_t ***table, uint64_t *len);
